@@ -4,9 +4,9 @@
 // (c) 2024 www.ebenmonney.com/mit-license
 // ---------------------------------------
 
-import { Component, OnInit, ViewChild, Input, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, ViewChild, Input, Output, EventEmitter, inject } from '@angular/core';
 import { HttpErrorResponse } from '@angular/common/http';
-import { NgModel, NgForm } from '@angular/forms';
+import { NgModel, NgForm, FormsModule } from '@angular/forms';
 
 import { AlertService, MessageSeverity } from '../../services/alert.service';
 import { AccountService } from '../../services/account.service';
@@ -15,13 +15,23 @@ import { User } from '../../models/user.model';
 import { UserEdit } from '../../models/user-edit.model';
 import { Role } from '../../models/role.model';
 import { Permissions } from '../../models/permission.model';
+import { NgClass } from '@angular/common';
+import { AutofocusDirective } from '../../directives/autofocus.directive';
+import { EqualValidator } from '../../directives/equal-validator.directive';
+import { NgSelectComponent, NgLabelTemplateDirective, NgOptionTemplateDirective } from '@ng-select/ng-select';
+import { TranslateModule } from '@ngx-translate/core';
 
 @Component({
-  selector: 'app-user-info',
-  templateUrl: './user-info.component.html',
-  styleUrl: './user-info.component.scss'
+    selector: 'app-user-info',
+    templateUrl: './user-info.component.html',
+    styleUrl: './user-info.component.scss',
+    standalone: true,
+    imports: [FormsModule, AutofocusDirective, NgClass, EqualValidator, NgSelectComponent, NgLabelTemplateDirective, NgOptionTemplateDirective, TranslateModule]
 })
 export class UserInfoComponent implements OnInit {
+  private alertService = inject(AlertService);
+  private accountService = inject(AccountService);
+
   public isEditMode = false;
   public isNewUser = false;
   public isSaving = false;
@@ -35,9 +45,9 @@ export class UserInfoComponent implements OnInit {
 
   public formResetToggle = true;
 
-  public changesSavedCallback: { (): void } | undefined;
-  public changesFailedCallback: { (): void } | undefined;
-  public changesCancelledCallback: { (): void } | undefined;
+  public changesSavedCallback: (() => void) | undefined;
+  public changesFailedCallback: (() => void) | undefined;
+  public changesCancelledCallback: (() => void) | undefined;
 
   @Input()
   isViewOnly = false;
@@ -73,10 +83,6 @@ export class UserInfoComponent implements OnInit {
 
   @ViewChild('roles')
   public roles!: NgModel;
-
-
-  constructor(private alertService: AlertService, private accountService: AccountService) {
-  }
 
   ngOnInit() {
     if (!this.isGeneralEditor) {
