@@ -6,12 +6,16 @@
 
 import { Component } from '@angular/core';
 import { fadeInOut } from '../../services/animations';
+import { RouterLink } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
 
 @Component({
-  selector: 'app-not-found',
-  templateUrl: './not-found.component.html',
-  styleUrl: './not-found.component.scss',
-  animations: [fadeInOut]
+    selector: 'app-not-found',
+    templateUrl: './not-found.component.html',
+    styleUrl: './not-found.component.scss',
+    animations: [fadeInOut],
+    standalone: true,
+    imports: [RouterLink, TranslateModule]
 })
 export class NotFoundComponent {
 }
