@@ -4,84 +4,69 @@
 // (c) 2024 www.ebenmonney.com/mit-license
 // ---------------------------------------
 
-import { Component, OnInit, OnDestroy, ViewChild } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
+import { Component, inject, OnInit, OnDestroy, AfterViewInit } from '@angular/core';
+import { ActivatedRoute, Router, RouterLink } from '@angular/router';
+
+import { TranslateModule } from '@ngx-translate/core';
+import { NgbNavModule } from '@ng-bootstrap/ng-bootstrap';
 import { Subscription } from 'rxjs';
 
 import { fadeInOut } from '../../services/animations';
-import { BootstrapTabDirective, EventArg } from '../../directives/bootstrap-tab.directive';
 import { AccountService } from '../../services/account.service';
 import { Permissions } from '../../models/permission.model';
+
+import { UserInfoComponent } from '../controls/user-info.component';
+import { UserPreferencesComponent } from '../controls/user-preferences.component';
+import { UsersManagementComponent } from '../controls/users-management.component';
+import { RolesManagementComponent } from '../controls/roles-management.component';
 
 
 @Component({
   selector: 'app-settings',
   templateUrl: './settings.component.html',
   styleUrl: './settings.component.scss',
-  animations: [fadeInOut]
+  animations: [fadeInOut],
+  standalone: true,
+  imports: [
+    RouterLink, TranslateModule, NgbNavModule,
+    UserInfoComponent, UserPreferencesComponent, UsersManagementComponent, RolesManagementComponent
+  ]
 })
-export class SettingsComponent implements OnInit, OnDestroy {
-  isProfileActivated = true;
-  isPreferencesActivated = false;
-  isUsersActivated = false;
-  isRolesActivated = false;
-
-  fragmentSubscription: Subscription | undefined;
+export class SettingsComponent implements OnInit, AfterViewInit, OnDestroy {
+  private router = inject(Router);
+  public route = inject(ActivatedRoute);
+  private accountService = inject(AccountService);
 
   readonly profileTab = 'profile';
   readonly preferencesTab = 'preferences';
   readonly usersTab = 'users';
   readonly rolesTab = 'roles';
-
-  @ViewChild('tab', { static: true })
-  tab!: BootstrapTabDirective;
-
-
-  constructor(private router: Router, private route: ActivatedRoute, private accountService: AccountService) {
-  }
+  activeTab = '';
+  showDatatable = false; // Delays showing the table until tab is shown so column widths are calculated correctly
+  fragmentSubscription: Subscription | undefined;
 
   ngOnInit() {
-    this.fragmentSubscription = this.route.fragment.subscribe(anchor => this.showContent(anchor));
+    this.fragmentSubscription = this.route.fragment.subscribe(fragment => this.setActiveTab(fragment));
+  }
+
+  ngAfterViewInit() {
+    setTimeout(() => this.showDatatable = true);
   }
 
   ngOnDestroy() {
     this.fragmentSubscription?.unsubscribe();
   }
 
-  showContent(anchor: string | null) {
-    if (anchor) {
-      anchor = anchor.toLowerCase();
-    }
+  setActiveTab(fragment: string | null) {
+    fragment = fragment?.toLowerCase() ?? this.profileTab;
 
-    if ((this.isFragmentEquals(anchor, this.usersTab) && !this.canViewUsers) ||
-      (this.isFragmentEquals(anchor, this.rolesTab) && !this.canViewRoles)) {
-      return;
-    }
+    const canViewTab = fragment === this.profileTab || fragment === this.preferencesTab ||
+      (fragment === this.usersTab && this.canViewUsers) || (fragment === this.rolesTab && this.canViewRoles);
 
-    this.tab.show(`#${anchor || this.profileTab}Tab`);
-  }
-
-  isFragmentEquals(fragment1: string | null, fragment2: string | null) {
-    if (fragment1 == null) {
-      fragment1 = '';
-    }
-
-    if (fragment2 == null) {
-      fragment2 = '';
-    }
-
-    return fragment1.toLowerCase() === fragment2.toLowerCase();
-  }
-
-  onShowTab(event: EventArg) {
-    const activeTab = (event.target as HTMLAnchorElement).hash.split('#', 2).pop();
-
-    this.isProfileActivated = activeTab === this.profileTab;
-    this.isPreferencesActivated = activeTab === this.preferencesTab;
-    this.isUsersActivated = activeTab === this.usersTab;
-    this.isRolesActivated = activeTab === this.rolesTab;
-
-    this.router.navigate([], { fragment: activeTab });
+    if (canViewTab)
+      this.activeTab = fragment;
+    else
+      this.router.navigate([], { fragment: this.profileTab });
   }
 
   get canViewUsers() {
