@@ -19,7 +19,7 @@ export class AlertService {
 
   showDialog(message: string): void;
   showDialog(message: string, type: DialogType, okCallback: (val?: string) => void): void;
-  showDialog(message: string, type: DialogType, okCallback?: { (val?: string): void } | null, cancelCallback?: { (): void } | null, okLabel?: string | null, cancelLabel?: string | null, defaultValue?: string | null): void;
+  showDialog(message: string, type: DialogType, okCallback?: ((val?: string) => void) | null, cancelCallback?: (() => void) | null, okLabel?: string | null, cancelLabel?: string | null, defaultValue?: string | null): void;
   showDialog(message: string, type?: DialogType, okCallback?: (val?: string) => void, cancelCallback?: () => void, okLabel?: string, cancelLabel?: string, defaultValue?: string) {
     if (!type) {
       type = DialogType.alert;
